Refactoring of dto
