+++ API/Interfaces/IUserInfoService.cs
@@ -1,10 +1,11 @@
-﻿using DomainModels.Models;
+﻿using DomainModels.Dto.UserProfileDto;
+using DomainModels.Models;
 
 namespace API.Interfaces;
 
 public interface IUserInfoService
 {
-	//Create and delete are not needed for UserInfo, as it is tied to User.
-	Task<UserInfo?> GetByUserIdAsync(int userId);
-	Task<UserInfo?> UpdateUserInfoAsync(int userId, UserInfo updatedInfo);
+	Task<UserInfoGetDto?> GetByUserEmailAsync(int id);
+	Task<UserInfoPutDto?> UpdateUserInfoAsync(int id, UserInfo updatedInfo);
+	Task<UserInfoPostDto?> CreateUserInfoAsync(UserInfo newInfo);
 }
