+++ API/Program.cs
@@ -35,7 +35,8 @@ public class Program
 		builder.Services.AddScoped<HotelService>();
 		builder.Services.AddScoped<IBookingInterface, BookingService>();
         builder.Services.AddScoped<ILoginAttemptService, LoginAttemptService>();
-        builder.Services.AddMemoryCache();
+		builder.Services.AddScoped<IUserInfoService, UserInfoService>();
+		builder.Services.AddMemoryCache();
 
         // Learn more about configuring OpenAPI at https://aka.ms/aspnet/openapi
         builder.Services.AddSwaggerGen(c =>
