+++ DomainModels/Dto/UserProfileDto/UserInfoPutDto.cs
@@ -1,0 +1,16 @@
+﻿
+namespace DomainModels.Dto.UserProfileDto;
+
+public class UserInfoPutDto
+{
+	public int UserId { get; set; }
+	public string? FirstName { get; set; }
+	public string? LastName { get; set; }
+	public string? Address { get; set; } 
+	public string? PostalCode { get; set; }
+	public string? City { get; set; }
+	public string? Country { get; set; }
+	public string? PhoneNumber { get; set; }
+	public DateOnly? DateOfBirth { get; set; }
+	public string? SpecialRequests { get; set; }
+}
